Fixes linked to the remove of enums more to come

+++ FlightStreamDeck.SimConnectFSX/Simvar.cs
@@ -12,6 +12,6 @@ namespace FlightStreamDeck.SimConnectFSX
 
 		public string Name { get; set; }
 
-		public float Data { get; set; }
+		public double Data { get; set; }
 	}
 }
